backend/requirements.txt
# Generated: 2025-03-10 — production dependencies only
flask>=3.0,<4
flask-cors>=4.0,<5
flask-sqlalchemy>=3.1,<4
flask-jwt-extended>=4.6,<5
flask-limiter>=3.5,<4
flask-migrate>=4.0,<5
flask-wtf>=1.2,<2
flask-mail>=0.10,<1
werkzeug>=3.0,<4
python-dotenv>=1.0,<2
markdown>=3.5,<4
bleach>=6.0,<7
gunicorn>=21.0,<23
